requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
alembic>=1.7.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=10.0
python-multipart>=0.0.5
aiohttp>=3.8.1
typer>=0.9.0
